# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API IT WORKS",
        "description": "Tài liệu API cho hệ thống ứng dụng IT WORKS",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/freelancer/info": {
            "get": {
                "tags": [
                    "Freelancer"
                ],
                "summary": "Get freelancer's information",
                "description": "Get freelancer's information.",
                "operationId": "getFreelancerInfo",
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns the freelancer information."
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to perform this action."
                    }
                }
            }
        },
        "/api/freelancer/info/update": {
            "post": {
                "tags": [
                    "Freelancer"
                ],
                "summary": "Update freelancer's information",
                "description": "Update freelancer's information.",
                "operationId": "updateFreelancerInfo",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns the updated freelancer information."
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to perform this action."
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error."
                    }
                }
            }
        },
        "/api/freelancer/mylist/{id}": {
            "post": {
                "tags": [
                    "Freelancer"
                ],
                "summary": "Add job to freelancer's list",
                "description": "Add a job to freelancer's list.",
                "operationId": "addJobToList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job to add to the list",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns success message.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Đã thêm thành công vào list của bạn."
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
        "/api/freelancer/mylist": {
            "get": {
                "tags": [
                    "Freelancer"
                ],
                "summary": "Get freelancer's job list",
                "description": "Get the list of jobs associated with the freelancer.",
                "operationId": "getFreelancerJobList",
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns the list of jobs."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}